cql: Refuse to create counter tables with tablets
The logic for counter updates does not handle intra-node tablet
migration. We may have more than one counter-owning replica on a given
host. During some migration stages, both replica shards should receive
the write. This is not implemented yet, and it's not trivial to do
because current locking wouldn't guarantee atomicity as it is
shard-local.

We decided to not support counters with tablets for now.