sstables: Fix incremental_reader_selector (for range reads) with tablets
incremental_reader_selector is the mechanism for incremental comsumption
of disjoint sstables on range reads.

tablet_sstable_set was implemented, such that incremental_reader_selector
works with tablets. But turns out the selector is incorrectly reading
tablet 1 even when asked to consume only tablet 0.

The reason is selector is vnode addicted and expects a given sstable
set's selector to return next_token_to_be_selected or max_token when
fully exhausted.
With tablets, when done selecting tablet 0, first_token of tablet 1
becomes the next_token_to_be_selected, and today i__r__selector
unconditionally selects tablet 1 even when range read targets
tablet 0 only. It wants to go all the way to max_token, that's why
it's vnode addicted.

So all tablets, in a given shard, are selected even when reading
from a single tablet. The selector doesn't know when to stop.
It doesn't cause a correctness issue as sstable reader will be created
with the range of the read, but it becomes an efficiency one as
the readers are created and fed into combined reader that will try
to consume from them, and then it'll have to discard them once it
realizes the sstables (from different tablets) have no data for
that range.

That can be seen from logs:
```
table - make_reader_v2 - range=(-inf, {-4611686018427387905, end}]
    incremental_reader_selector - create_new_readers(null): selecting on pos {minimum token, w=-1}
    sstable - make_reader - reader on (-inf, {-4611686018427387905, end}] for sst 3gfx_..._34qn42... that has range [{-9151620220812943033, start},{-4813568684827439727, end}]
    incremental_reader_selector - create_new_readers(null): selecting on pos {-4611686018427387904, w=-1}
    sstable - make_reader - reader on (-inf, {-4611686018427387905, end}] for sst 3gfx_..._368nk2... that has range [{-4599560452460784857, start},{-78043747517466964, end}]
    incremental_reader_selector - create_new_readers(null): selecting on pos {0, w=-1}
    sstable - make_reader - reader on (-inf, {-4611686018427387905, end}] for sst 3gfx_..._38lj42... that has range [{851021166589397842, start},{3516631334339266977, end}]
    incremental_reader_selector - create_new_readers(null): selecting on pos {4611686018427387904, w=-1}
    sstable - make_reader - reader on (-inf, {-4611686018427387905, end}] for sst 3gfx_..._3dba82... that has range [{5065088566032249228, start},{9215673076482556375, end}]
```

Fix is about making sure i__r__selector won't select past the
supplied range for the read.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>