service/tablet_allocator: Introduce tablet_candidate
This new type will allow the load balancer to treat co-located tablets
as a single candidate (will treat them as if they were already merged),
allowing co-located tablets to be migrated together (in the same
migration plan).

The type is a variant of global_tablet_id and colocated_tablet (which
holds the global_tablet_id of sibling tablets).

It will be eventually wired after some more preparation. It will allow
for minimal amount of changes in the balancer code.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>